#9 Implement: product update/delete api 연결, top button design 적용 Fix: detail image dynamic inflate로 변경
1. product update/delete api 연결, 결과 노티피케이션 미구현
2. NestedScrollView와 RecyclerView가 겹치면서 appbar밑으로 레이아웃이 들어가버리는 버그때문에 recyclerView 대신 dynamic inflate로 수정
3. top button에 radius적용된 design 적용

+++ app/src/main/java/kr/co/mash_up/a9tique/data/remote/BackendService.java
@@ -55,7 +55,6 @@ public interface BackendService {
     @POST("api/product")
     Observable<JsonObject> addProduct(@Body RequestProduct requestProduct);
 
-    /****************** Product ********************/
     /**
      * 상품 정보 삭제
      *
@@ -63,7 +62,7 @@ public interface BackendService {
      * @return 성공 or 실패
      */
     @DELETE("api/product/{id}")
-    Observable<JsonObject> deleteProduct(@Path("id") int productId);
+    Observable<JsonObject> deleteProduct(@Path("id") long productId);
 
     /**
      * 상품 정보 수정
@@ -73,7 +72,7 @@ public interface BackendService {
      * @return 성공 or 실패
      */
     @PUT("api/product/{id}")
-    Observable<JsonObject> updateProduct(@Path("id") int productId,
+    Observable<JsonObject> updateProduct(@Path("id") long productId,
                                          @Body RequestProduct requestProduct);
 
     /**
